requirements.txt
flask
python-dotenv
pandas
pdfplumber
openai
openpyxl
requests